attached_assets/Pasted-Error-React-jsx-type-is-invalid-expected-a-string-for-built-in-components-or-a-class-function-1735489460180.txt

Error: React.jsx: type is invalid -- expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.
    at createUnhandledError (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/client/components/react-dev-overlay/internal/helpers/console-error.js:27:49)
    at handleClientError (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/client/components/react-dev-overlay/internal/helpers/use-error-handler.js:44:56)
    at console.error (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/client/components/globals/intercept-console-error.js:48:56)
    at jsxDEVImpl (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react/cjs/react-jsx-dev-runtime.development.js:479:17)
    at exports.jsxDEV (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react/cjs/react-jsx-dev-runtime.development.js:658:14)
    at UploadPage (webpack-internal:///(app-pages-browser)/./src/app/upload/page.tsx:18:94)
    at react-stack-bottom-frame (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:23610:20)
    at renderWithHooks (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:4646:22)
    at updateFunctionComponent (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:8032:19)
    at beginWork (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:9689:18)
    at runWithFiberInDEV (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:544:16)
    at performUnitOfWork (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:15064:22)
    at workLoopSync (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14888:41)
    at renderRootSync (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14868:11)
    at performWorkOnRoot (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:14394:44)
    at performWorkOnRootViaSchedulerTask (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/react-dom/cjs/react-dom-client.development.js:15955:7)
    at MessagePort.performWorkUntilDeadline (webpack-internal:///(app-pages-browser)/../node_modules/next/dist/compiled/scheduler/cjs/scheduler.development.js:44:48)
